Full changelog:
 Added Award
 Added IAwardRepository
 Added AwardRepository
 Added Banned User
 Added IBannedUserRepository
 Added BannedUserRepository
 Added Comment
 Added ICommentRepository
 Added CommentRepository
 Added Post
 Added IPostRepository
 Added PostRepository
 Added Report
 Added IReportRepository
 Added ReportRepository
 Added Vote
 Added IVoteRepository
 Added VoteRepository
 Added VoteableEntity
 Changed post, comment, award to use VoteableEntity instead of Entity.

VoteableEntity

VoteableEntity represents an entity that can be voted, awarded, or reported. The base object also contains data on archive and deletion. This will help with defining entities that should be given an award, or can be commented on. Entity framework has 3 lists it will populate based on the child objects ID.

2 objects contain this new class: Post, Comment,